Fix variable names containing ignorable chars not being renamed
Ignorable chars (Character.isIdentifierIgnorable) are ignored as part of
identifiers (cannot be leading, but can be in between and trailing) when
compiling. They can therefore normally not appear in a class file unless
it was modified. Since they cannot be represented in source (and could
easily cause identifier collisions), identifiers containing them are now
considered invalid.

This was only added to the JLS for Java 10 (see JDK-6214519), however the
behavior also exists in earlier versions.